% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplineImportance.R
\name{bsplineImportance}
\alias{bsplineImportance}
\title{Correlation based variable importance weights using B-splines for continuous
predictors to find non-linear effects}
\usage{
bsplineImportance(formula, data, df = 5)
}
\arguments{
\item{formula}{an R formula to specify the categorical target to match and the
numeric or integer predictors}

\item{data}{a data frame (or matrix that will be coerced to data frame) that
contains the variables specified in the formula}

\item{df}{the degrees for freedom to used in the B-spline function of the
potential predictors, with a default set at five}
}
\value{
The function returns a data frame with three columns, the first is
the variable name, the second is its importance weight, and the third is the
linearity measure for continuous predictors, NA for categorical
predictors.
}
\description{
\code{bsplineImportance} returns a data frame of variable importance
weights and a measure of monotonicity for a set of predictor variables
for a continuous target variable.
}
\details{
Determine the strength of the relationship between a target and a potential
predictor, based on the correlation between fitted actual values of a linear
regression between the B-spline transformed predictor and the target, and
whether the relationship appears to be linear based on the ratio of the
Pearson correlation coeficient between the a potential predictor and the
target over the importance measure. This measure is bounded between zero and
and one, with higher values indicating a more linear relationship.
}

