% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_int_to_factor.R, R/cols_factor_is_number.R
\name{change_factor_to_number}
\alias{change_factor_to_number}
\title{Change factor type columns to numeric.}
\usage{
change_factor_to_number(df, factor.cols)

change_factor_to_number(df, factor.cols)
}
\arguments{
\item{df}{The data.frame}

\item{factor.cols}{A vector of the names of the factor columns to be
converted to numerics}

\item{fact.cols}{A vector of the names of the integer columns to be
converted to factors}

\item{df}{The data.frame}
}
\value{
The modified data.frame.

The modified data.frame.
}
\description{
Change factor type columns to numeric.

Change factor type of columns to numeric type.
}

