% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{change_float_to_int}
\alias{change_float_to_int}
\title{Change float column to int, if it essentially is an integer type}
\usage{
change_float_to_int(df, float_cols)
}
\arguments{
\item{df}{The data.frame}

\item{float_cols}{The columns that is of float type that should be potentially changed}
}
\value{
Modified data.frame
}
\description{
Change float column to int, if it essentially is an integer type
}

