% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_int_to_factor.R
\name{change_int_to_factor}
\alias{change_int_to_factor}
\title{Change integer type of columns to factor type.}
\usage{
change_int_to_factor(df, int.factor.cols)
}
\arguments{
\item{df}{The data.frame}

\item{int.factor.cols}{A vector of the names of the integer columns to be
converted to factors}
}
\value{
The modified data.frame.
}
\description{
Change integer type of columns to factor type.
}

