% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{change_missing_to_na}
\alias{change_missing_to_na}
\title{Convert whitespace entries to NAs}
\usage{
change_missing_to_na(df)
}
\arguments{
\item{df}{The data.frame you wish to change missing values to NAs.}
}
\description{
Convert whitespace entries to NAs
}

