% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_numeric_to_int.R
\name{change_numeric_to_int}
\alias{change_numeric_to_int}
\title{Change a numeric column to integer, if it appears to be an integer}
\usage{
change_numeric_to_int(df, numeric.int.cols)
}
\arguments{
\item{df}{The data frame}

\item{numeric.int.cols}{Columns that are numeric but shoud be integers}
}
\value{
Modified data.frame
}
\description{
Change a numeric column to integer, if it appears to be an integer
}

