% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquaredImportance.R
\name{chiSquaredImportance}
\alias{chiSquaredImportance}
\title{Chi-squared statistic based variable importance weights}
\usage{
chiSquaredImportance(formula, data)
}
\arguments{
\item{formula}{an R formula to specify the categorical target to match and the
numeric or integer predictors}

\item{data}{a data frame (or matrix that will be coereced to a data frame)
that contains the variables specified in the formula}
}
\value{
The function returns a data frame with three columns: the variable
names, the importance weights, and (for binary targets) a linearity measure.
The rows of the data frame are sorted by the variables' importance weights.
}
\description{
\code{chiSquaredImportance} returns a data frame of variable importance
weights for a set of predictor variables for a factor target variable.
}
\details{
The potential predictors can be factor, integer, or numeric variables.
Integer and numeric variables are binned into factors using an entropy
based algorithm to find the best set of bins relative to the target
variable. The algorithm makes use of a minimum distance length stopping
rule determine the number of bins to use.
}

