% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_collinear.R
\name{cols_collinear}
\alias{cols_collinear}
\title{Groups of collinear columns within a data frame}
\usage{
cols_collinear(df)
}
\arguments{
\item{df}{A data.frame that may contain both number (numeric and integer)
and factor columns}
}
\value{
A named list each element of which is a character vector containing
the names of the collinear columns.
}
\description{
\code{cols_collinear} returns a named list with the columns that
represent a collinear group.
}
\details{
The collinear column groups are determined in a three stepp process. In the
first, model.matrix is used to convert the data from a data.frame to a
matrix where the factors have been converted to a complete set of zero-one
indicator variables, and its cross product is calculated. In the second
step, singular value decomposition of the cross product in performed to find
the total number of collinear groups in the data is found using the number
of singular values that are smaller than 1e-7. In the third step the columns
of the V matrix from the singular value decompostion that correspond to the
very small singular values are subjected to a varimax rotation in order to
identify which columns in the data cause the small singular values to occur.
The collinear groups are then pruned to remove the trival cases associated
with all of the factor levels in the data being placed into a single group
and groups whose members correspond to a strict subset of the columns
contained in another group.
}

