% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_duplicates.R
\name{cols_continuous_duplicates}
\alias{cols_continuous_duplicates}
\title{Find duplicate and perfectly correlated numeric and integer columns in a
data frame}
\usage{
cols_continuous_duplicates(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{cols_continuous_duplicates} returns a named list, each entry of which is
the the set of other columns that represent a duplicate of, or are perfectly
correlated with, the named column.
}
\details{
Duplicates and perfectly correlated columns are found using the absolute
value of pairwise Pearson correlation coefficients. As a result, two
that perfectly negatively correlated will be included in the set of returned
columns. The pairwise comparisons are made with missing values removed,
so the comparison is really being made for cases where neither variable has
missing values.
}
\seealso{
\code{\link{cols_duplicates}}, \code{\link{cols_factor_duplicates}}
}

