% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_continuous_fitness.R
\name{cols_continuous_fitness}
\alias{cols_continuous_fitness}
\title{Calculate the fitness measure for continuous columns.}
\usage{
cols_continuous_fitness(cm)
}
\arguments{
\item{cm}{The data.frame of continuous column fitness metrics}
}
\value{
A named vector of fitness measures that can range from zero to one.
}
\description{
Calculate the fitness measure for continuous columns.
}

