% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_duplicates.R
\name{cols_duplicates}
\alias{cols_duplicates}
\title{Find duplicate and perfectly correlated integer columns in a data frame}
\usage{
cols_duplicates(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{cols_duplicates} returns a named list, each entry of which is
the the set of other columns that represent a duplicate of, or are perfectly
correlated with, the named column.
}
\details{
The pairwise comparisons are made with missing values removed,so the
comparison is really being made for cases where neither variable has
missing values. Details of the methods used can be found in the help pages
for \link{cols_factors_duplicates} (in the case of factor columns) and
\link{cols_factors_duplicates} (in the case of integer and factor columns).
}
\seealso{
\code{\link{cols_continuous_duplicates}},
\code{\link{cols_factors_duplicates}}
}

