% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_duplicates.R
\name{cols_factor_duplicates}
\alias{cols_factor_duplicates}
\title{Find duplicate factor columns in a data frame}
\usage{
cols_factor_duplicates(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
\code{cols_factor_duplicates} returns a named list, each entry of which is
the the set of other columns that represent a duplicate of the named column.
}
\details{
Duplicates are found using Goodman-Kurskal tau statistic. Two variables
(x1 and x2) are considered to be duplicates if the tau statistic of x1 for
x2 is one, and the tau statistic of x2 for x1 is also one. The pairwise
comparisons are made with missing values removed, so the comparison is really
being made for cases where neither variable has missing values.
}
\seealso{
\code{\link{cols_duplicates}}, \code{\link{cols_continuous_duplicates}}
}

