% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_factor_fitness.R
\name{cols_factor_fitness}
\alias{cols_factor_fitness}
\title{Calculate the fitness measure for factor columns.}
\usage{
cols_factor_fitness(fm)
}
\arguments{
\item{fm}{The data.frame of factor column fitness metrics}
}
\value{
A named vector of fitness measures that can range from zero to one.
}
\description{
Calculate the fitness measure for factor columns.
}

