% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_factor_is_number.R
\name{cols_factor_is_number}
\alias{cols_factor_is_number}
\title{Determine which factors might actually be numbers.}
\usage{
cols_factor_is_number(df, factor.cols)
}
\arguments{
\item{df}{a data frame}

\item{factor.cols}{a vector containing the names of the factor variables
of interest}
}
\description{
Determine which factors might actually be numbers.
}

