% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_imputation.R
\name{cols_imputation}
\alias{cols_imputation}
\title{Impute missing values, with factors getting the value Missing, and numerics
and integers getting zero (strictly positive valued columns) or a negative
value that is less than the existing minimum value. An imputation flag is
also created for numeric and integer columns.}
\usage{
cols_imputation(df)
}
\arguments{
\item{df}{The data.frame}
}
\value{
The modified data.frame.
}
\description{
Impute missing values, with factors getting the value Missing, and numerics
and integers getting zero (strictly positive valued columns) or a negative
value that is less than the existing minimum value. An imputation flag is
also created for numeric and integer columns.
}

