% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_int_is_categorical}
\alias{cols_int_is_categorical}
\title{Return column names, whose current data type is integer, but it is potentially
categorical.}
\usage{
cols_int_is_categorical(df, int_cols, threshold = 2)
}
\arguments{
\item{df}{The data.frame}

\item{int_cols}{A vector of strings, the column nams of integer columns.}

\item{threshold}{An integer or numeric number. If the number of unique values
in the column is less than or equal to threshold, the column names will be returned.}
}
\value{
Names of columns who are integer type but are potentially categorical.
}
\description{
Return column names, whose current data type is integer, but it is potentially
categorical.
}

