% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_int_sequential.R
\name{cols_int_sequential}
\alias{cols_int_sequential}
\title{Determine which integer columns have sequential values}
\usage{
cols_int_sequential(df, integer.cols)
}
\arguments{
\item{df}{a data frame or a matrix}

\item{integer.cols}{a character vector containing the names of integer variables}
}
\description{
Determine which integer columns have sequential values
}

