% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_integer_is_numeic.R
\name{cols_integer_is_numeric}
\alias{cols_integer_is_numeric}
\title{Determine which numeric variables in a data frame are actually integers}
\usage{
cols_integer_is_numeric(df, numeric.cols)
}
\arguments{
\item{df}{a data frame or a matrix}

\item{numeric.cols}{a character vector containing the names of numeric variables}
}
\description{
Determine which numeric variables in a data frame are actually integers
}

