% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_missing_blank.R
\name{cols_missing_blank}
\alias{cols_missing_blank}
\title{Report on the number and percentage of missing (NA) values for a data.frame}
\usage{
cols_missing_blank(df)
}
\arguments{
\item{df}{The data.frame}
}
\value{
A data.frame containing the missing value report.
}
\description{
Report on the number and percentage of missing (NA) values for a data.frame
}

