% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_num_outliers.R
\name{cols_num_outliers}
\alias{cols_num_outliers}
\title{Conservatively detect outliers for continuous variables in a data frame}
\usage{
cols_num_outliers(df, num.cols, method = match.arg(c("boxplot", "3sigma",
  "4sigma")))
}
\arguments{
\item{df}{A data.frame}

\item{method}{A string, specifying the .It can
be 'quantile' or '3sigma (default)'.}

\item{num.col}{A vector of the names of the numeric variables in the
data.frame}

\item{IQR_coef}{Coefficient of IQR (internal quantile range) in box plot.
This parameter is optional and only used for strategy "quantile".The
suggested value for this parameter is between 1.5 (default) to 3, with 1.5
less conservative and returning all suspected outliers and 3 only returning
values the algorithm is more certain to be outliers.}

\item{return_stats}{If TRUE, the function returns a list with information of
if there are outliers, number of outlier, percentage of outlier in the
column. If FALSE, the function returns the records themselves who are
outliers.}
}
\value{
A boolean to indicate if there are outliers or records that are
outliers
}
\description{
Conservatively detect outliers for continuous variables in a data frame
}

