% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_num_sparse_levels.R
\name{cols_num_sparse_levels}
\alias{cols_num_sparse_levels}
\title{Convert any blank values to missing (NA) values}
\usage{
cols_num_sparse_levels(df, cols.to.use)
}
\arguments{
\item{df}{The data.frame}

\item{cols.to.use}{The vector of column names for which sparse levels are 
reported}
}
\value{
A data.frame containing the sparse levels report.
}
\description{
Convert any blank values to missing (NA) values
}

