% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_numeric_is_integer.R
\name{cols_numeric_is_integer}
\alias{cols_numeric_is_integer}
\title{Determine which numeric fields might actually be integers}
\usage{
cols_numeric_is_integer(df, numeric.cols)
}
\arguments{
\item{df}{a data frame}

\item{numeric.cols}{a vector containing the names of the numeric varaibles
of interest}
}
\description{
Determine which numeric fields might actually be integers
}

