% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_numeric_outliers}
\alias{cols_numeric_outliers}
\title{Return the names of columns who have outliers.}
\usage{
cols_numeric_outliers(df, num_cols, strategy = "quantile")
}
\arguments{
\item{df}{The data.frame}

\item{num_cols}{A vector of string, the names of numeric columns you wish to
identify if they have outliers.}

\item{strategy}{A string, which strategy to use to identify outlier.It can
to 'quantile'(default) or '3sigma'.}
}
\value{
Names of numeric columns who are likely to have outliers
}
\description{
Return the names of columns who have outliers.
}

