% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_possible_id}
\alias{cols_possible_id}
\title{Return names of columns that are possibly ID columns}
\usage{
cols_possible_id(df, cols, threshold = 0.95)
}
\arguments{
\item{df}{The data.frame to examine}

\item{cols}{The names of columns to examine}

\item{threshold}{The threshold of possible id by percentage of unique
numbers, the default is 0.95}
}
\value{
Names of columns that are likely to be ID columns
}
\description{
Return names of columns that are possibly ID columns
}

