% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_possible_nums}
\alias{cols_possible_nums}
\title{Return the names of factor columns who are likely to be numeric.}
\usage{
cols_possible_nums(df, cat_cols, threshold = 5)
}
\arguments{
\item{df}{The data.frame}

\item{cat_cols}{A vector of strings, names of categorical columns}

\item{threshold}{An integer or numeric number. If there are more than
threshold or 10% number of certain patterns we know they are numeric,
return the column names.}
}
\value{
Names of factor type columns who are likely to be numerical.
}
\description{
Return the names of factor columns who are likely to be numeric.
}

