% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_sparse_categorical}
\alias{cols_sparse_categorical}
\title{Return names of categorical columns with very sparse levels or very unbalanced.}
\usage{
cols_sparse_categorical(df, fact_cols, freqCut = 19, numberCut = 10)
}
\arguments{
\item{df}{The data.frame to examine}

\item{fact_cols}{The names of factor columns to examine}

\item{freqCut}{The threshod of the ratio of the most frequent level devided
by the second most frequent level}

\item{numberCut}{The threshold of the number of least frequent level.}
}
\value{
Names of columns that has too sparse or unbalanced levels
}
\description{
Return names of categorical columns with very sparse levels or very unbalanced.
}

