% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_high_hhi}
\alias{cols_with_high_hhi}
\title{Return names of columns with high Herfindahl-Hirschman index. This is only
for categorical predictors.}
\usage{
cols_with_high_hhi(df, fact_cols, threshold = 9000)
}
\arguments{
\item{df}{The data.frame to examine}

\item{fact_cols}{The names of factor columns to examine}

\item{threhold}{The HHI threshold. Note that HHI is in the range (0, 10000].
The higher the HHI, the more monopoly some levles have.}
}
\value{
A vector of column names with high HHI
}
\description{
Return names of columns with high Herfindahl-Hirschman index. This is only
for categorical predictors.
}

