% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_low_sd}
\alias{cols_with_low_sd}
\title{Return names of columns who have low standard deviation}
\usage{
cols_with_low_sd(df, num_cols = NULL, threshold = 0.01, scale = FALSE)
}
\arguments{
\item{df}{The data.frame to examine}

\item{num_cols}{The name of columns of numeric type to examine}

\item{threshold}{The threshold of stardard deviation that is considered as
too low.}

\item{scale}{can be FALSE, 'zeroOne' or 'IQR'. If scale is 'zeroOne', the
data to the interval of [0,1] before analyzing its standard deviation; if
scale is 'IQR', it returns standard deviation divided by (Q3 - Q1)}
}
\value{
A vector of column names with low standard deviation
}
\description{
Return names of columns who have low standard deviation
}

