% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_many_levels}
\alias{cols_with_many_levels}
\title{Return names of categorical columns with too many levels to be useful.}
\usage{
cols_with_many_levels(df, cat_cols = NULL, thres_percent = 0.5,
  thres_absolute = 100)
}
\arguments{
\item{df}{The data.frame to examine}

\item{cat_cols}{The names of categorical columns to examine}

\item{threshold}{The threshold of number of levels as percentage of record
number. The default is 50\%}
}
\value{
Names of categorical columns with too many levels
}
\description{
Return names of categorical columns with too many levels to be useful.
}

