% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_many_same}
\alias{cols_with_many_same}
\title{Return names of columns who has the same value over the threshold}
\usage{
cols_with_many_same(df, cat_cols = colnames(df), threshold = 0.95)
}
\arguments{
\item{df}{The data.frame to examine}

\item{cat_cols}{The name of columns you wish to examine}

\item{threshold}{The threshold of same value percentage in a column,
95\% by default}
}
\value{
A vector of column names who has many same values
}
\description{
Return names of columns who has the same value over the threshold
}

