% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_nzv}
\alias{cols_with_nzv}
\title{Return names of columns with near zero variance.}
\usage{
cols_with_nzv(df, num_cols = NULL, uniqueCut = 10, freqCut = 19)
}
\arguments{
\item{df}{The data.frame to examine}

\item{num_cols}{The names of columns to examine}

\item{uniqueCut}{Percentage of allowed unique values.}

\item{freqCut}{Ratio allowed of the most and the second most frequent values}
}
\value{
A vector of column names with near zero variance
}
\description{
Return names of columns with near zero variance.
}

