% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{cols_with_unique_val}
\alias{cols_with_unique_val}
\title{Return names of columns that has only one value}
\usage{
cols_with_unique_val(df, cols = NULL)
}
\arguments{
\item{df}{The data.frame to examine}

\item{cols}{Names of columns to examine, will use all columns of df, if it
is not specified.}
}
\value{
A vector of columns names with only one value
}
\description{
Return names of columns that has only one value
}

