% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{detect_outlier_single}
\alias{detect_outlier_single}
\title{Detect outliers for single variable}
\usage{
detect_outlier_single(s, strategy = "3sigma", IQR_coef = 1.5,
  return_records = TRUE)
}
\arguments{
\item{s}{One column of a data.frame}

\item{IQR_coef}{Coefficient of IQR (internal quantile range) in box plot.
This parameter is optional and only used for strategy "quantile".The
suggested value for this parameter is between 1.5 (default) to 3, with 1.5
less conservative and returning all suspected outliers and 3 only returning
values the algorithm is more certain to be outliers.}

\item{stategy}{A string, which strategy to use to identify outlier.It can
be 'quantile' or '3sigma (default)'.}

\item{return_stats}{If TRUE, the function returns a list with information of
if there are outliers, number of outlier, percentage of outlier in the
column. If FALSE, the function returns the records themselves who are
outliers.}
}
\value{
A boolean to indicate if there are outliers or records that are
outliers
}
\description{
Detect outliers for single variable
}

