% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyBins.R
\name{entropyBins}
\alias{entropyBins}
\title{Entropy target based bining}
\usage{
entropyBins(x, y, label.type = c("letters", "intervals"),
  na.as.missing = TRUE)
}
\arguments{
\item{x}{a numeric vector to be binned}

\item{y}{a factor vector to determine the split points}

\item{label.type}{indicates the type of label levels to use for the levels
in the created factor variable. The choices are lower case letters or
numeric intervals}

\item{na.as.missing}{a flag to indicate whether missing values should be a category
or remain as NA values}
}
\value{
The function returns a data frame with two columns, the first is
the variable name, the second is its importance weight. The rows of the data
frame are sorted by the variables' importance weights.
}
\description{
\code{entropyBins} finds the optimal bins for a numeric or integer variable
to match a factor variable based on entropy
}
\details{
The entropy bining algorithm makes use of a minimum distance length stopping
rule to determine the number of bins to use.
}

