% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropySplit.R
\name{entropySplit}
\alias{entropySplit}
\title{Finds the value of a numeric vector that results in the best split based on
a factor variable}
\usage{
entropySplit(x, y, ordered = FALSE)
}
\arguments{
\item{x}{a numeric vector to be split}

\item{y}{a factor vector to determine the split point}

\item{ordered}{a logical which indicates whether x and y have been ordered by the values of x}

\item{possible_splits}{a vector of indicies of possible split points, which is calculated if its value is NULL}
}
\description{
Finds the value of a numeric vector that results in the best split based on
a factor variable
}

