% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalIntervalBins.R
\name{equalIntervalBins}
\alias{equalIntervalBins}
\title{Bin a numeric variable into a set of equal interval bins.}
\usage{
equalIntervalBins(x, num.bins, na.as.missing = TRUE)
}
\arguments{
\item{x}{a numeric vector to be split}

\item{num.bins}{a length one integer vector specifying the number of bins to use}

\item{na.as.missing}{a flag to indicate whether missing values should be a category
or remain as NA values}
}
\description{
Bin a numeric variable into a set of equal interval bins.
}

