% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattedTextList.R
\name{formattedTextList}
\alias{formattedTextList}
\title{An English language centric helper function that converts a vector into a
comma seperated list of the values in the vector as a character string.}
\usage{
formattedTextList(x)
}
\arguments{
\item{x}{a vector of values to be converted into a text string list}
}
\description{
An English language centric helper function that converts a vector into a
comma seperated list of the values in the vector as a character string.
}

