% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{get_col_types}
\alias{get_col_types}
\title{Return names of columns that are factor, integer or numeric respectively}
\usage{
get_col_types(df)
}
\arguments{
\item{df}{The data.frame to examine}
}
\value{
A list of 4 string vectors of factor, integer and numeric and
everything else columns
}
\description{
Return names of columns that are factor, integer or numeric respectively
}

