% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataadviser.R
\name{get_HHI}
\alias{get_HHI}
\title{Calculate the Hefindahl Hirschman index for the levels of a factor}
\usage{
get_HHI(s)
}
\arguments{
\item{s}{A factor vector}
}
\description{
Calculate the Hefindahl Hirschman index for the levels of a factor
}

