% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyBins.R
\name{intervalLabels}
\alias{intervalLabels}
\title{A helper function to construct interval labels}
\usage{
intervalLabels(bdf)
}
\arguments{
\item{bdf}{a data frame containing information about the upper and lower
bounds of the categories being created from a numeric or integer column}
}
\description{
A helper function to construct interval labels
}

