% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posLog.R
\name{posLog}
\alias{posLog}
\title{Produces logarithm values for strictly positive values, zero otherwise}
\usage{
posLog(x, base = exp(1))
}
\arguments{
\item{x}{numeric vector}

\item{base}{of the logarithm, default is exp(1), the natural logarithm}
}
\description{
Produces logarithm values for strictly positive values, zero otherwise
}

