% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_factor_is_number.R
\name{strip_helper}
\alias{strip_helper}
\title{Strip off characters such as currency indicator and percentage signs that
cause what is actually a numeric or integer column to be interpreted as a
character or factor column.}
\usage{
strip_helper(x)
}
\arguments{
\item{x}{a character vector which is typically a column in a data frame.}
}
\description{
Strip off characters such as currency indicator and percentage signs that
cause what is actually a numeric or integer column to be interpreted as a
character or factor column.
}

