% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strongCor.R
\name{strongCor}
\alias{strongCor}
\title{Find strong bivariate correlations}
\usage{
strongCor(df, threshold = 0.9)
}
\arguments{
\item{df}{a data frame containing at least two integer or numeric columns}

\item{threshold}{the value of a correlation (in absolute value terms) that
is considered to be strong.}
}
\value{
A names list, with the name corresponding to each variable that has
a strong correlation with one or more other variables. Each named item
consists of a character vector of the other variables it has a strong
relationship with.
}
\description{
For each continuous variable in a data frame find the variables it is
strongly correlated with, where "strong" is user defined.
}

