% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimNewlines.R
\name{trimNewlines}
\alias{trimNewlines}
\title{Strip out newlines and white space from elements of a character vector.}
\usage{
trimNewlines(x)
}
\arguments{
\item{x}{a character vector to clean}
}
\description{
Strip out newlines and white space from elements of a character vector.
}

